requirements.txt
RDKit >= 2018.03
Pytorch >= 1.0
networkx